3 Reasoning Act Observe/lab/requirements.txt
google-generativeai>=0.5.0
requests>=2.32.0
